Field properties

ElectricFields.wavelength — Function

wavelength(f)

Return the carrier wavelength $\lambda$ of f.

ElectricFields.period — Function

period(f)

Return the carrier period time $T$ of f.

ElectricFields.frequency — Function

frequency(f)

Return the carrier frequency $f$ of the field f.

ElectricFields.max_frequency — Function

max_frequency(f)

Return the maximum carrier frequency $f$ in case of a composite field f.

ElectricFields.wavenumber — Function

wavenumber(f)

Return the carrier wavenumber $\nu$ of the field f.

ElectricFields.photon_energy — Function

photon_energy(f)

Return the carrier photon energy $\hbar\omega$ of the field f.

IntervalSets.duration — Function

duration(f)

Return the pulse duration $\tau$ of the field f.

ElectricFields.continuity — Function

continuity(f)

Return the pulse continuity, i.e. differentiability, of the field f.

ElectricFields.fluence — Function

fluence(f)

Compute the fluence of the field f, i.e.

\[\frac{1}{\hbar\omega} \int\diff{t} I(t),\]

where $I(t)$ is the intensity envelope of the pulse.

ElectricFields.intensity — Function

intensity(f)

Return the peak intensity of the field f.

ElectricFields.amplitude — Function

amplitude(f)

Return the peak amplitude of the field f.

ElectricFields.carrier — Function

carrier(f)

Return the carrier of the field f.

ElectricFields.envelope — Function

envelope(f)

Return the envelope of the field f.

ElectricFields.phase — Function

phase(c)

Return the phase of the carrier c.

ElectricFields.dimensions — Function

dimensions(f)

Return the number of dimensions of the field f. See also polarization.

Examples

julia> @field(F) do
           I₀ = 2.0
           T = 2.0
           σ = 3.0
           Tmax = 3.0
       end
Linearly polarized field with
  - I₀ = 2.0000e+00 au = 7.0188904e16 W cm⁻² =>
    - E₀ = 1.4142e+00 au = 727.2178 GV m⁻¹
    - A₀ = 0.4502 au
  – a Fixed carrier @ λ = 14.5033 nm (T = 48.3777 as, ω = 3.1416 Ha = 85.4871 eV)
  – and a Gaussian envelope of duration 170.8811 as (intensity FWHM; ±2.00σ)
  – Uₚ = 0.0507 Ha = 1.3785 eV => α = 0.1433 Bohr = 7.5826 pm

julia> dimensions(F)
1

julia> @field(F) do
           I₀ = 2.0
           T = 2.0
           σ = 3.0
           Tmax = 3.0
           ξ = 1.0
       end
Transversely polarized field with
  - I₀ = 2.0000e+00 au = 7.0188904e16 W cm⁻² =>
    - E₀ = 1.4142e+00 au = 727.2178 GV m⁻¹
    - A₀ = 0.4502 au
  – a Elliptical carrier with ξ = 1.00 (RCP) @ λ = 14.5033 nm (T = 48.3777 as, ω = 3.1416 Ha = 85.4871 eV)
  – and a Gaussian envelope of duration 170.8811 as (intensity FWHM; ±2.00σ)
  – Uₚ = 0.0507 Ha = 1.3785 eV => α = 0.1433 Bohr = 7.5826 pm

julia> dimensions(F)
3

ElectricFields.polarization — Function

polarization(f)

Return the polarization kind of the field f. See also dimensions.

Examples

julia> @field(F) do
           I₀ = 2.0
           T = 2.0
           σ = 3.0
           Tmax = 3.0
       end
Linearly polarized field with
  - I₀ = 2.0000e+00 au = 7.0188904e16 W cm⁻² =>
    - E₀ = 1.4142e+00 au = 727.2178 GV m⁻¹
    - A₀ = 0.4502 au
  – a Fixed carrier @ λ = 14.5033 nm (T = 48.3777 as, ω = 3.1416 Ha = 85.4871 eV)
  – and a Gaussian envelope of duration 170.8811 as (intensity FWHM; ±2.00σ)
  – Uₚ = 0.0507 Ha = 1.3785 eV => α = 0.1433 Bohr = 7.5826 pm

julia> polarization(F)
LinearPolarization()

julia> @field(F) do
           I₀ = 2.0
           T = 2.0
           σ = 3.0
           Tmax = 3.0
           ξ = 1.0
       end
Transversely polarized field with
  - I₀ = 2.0000e+00 au = 7.0188904e16 W cm⁻² =>
    - E₀ = 1.4142e+00 au = 727.2178 GV m⁻¹
    - A₀ = 0.4502 au
  – a Elliptical carrier with ξ = 1.00 (RCP) @ λ = 14.5033 nm (T = 48.3777 as, ω = 3.1416 Ha = 85.4871 eV)
  – and a Gaussian envelope of duration 170.8811 as (intensity FWHM; ±2.00σ)
  – Uₚ = 0.0507 Ha = 1.3785 eV => α = 0.1433 Bohr = 7.5826 pm

julia> polarization(F)
ArbitraryPolarization()

Strong-field properties

ElectricFields.ponderomotive_potential — Function

ponderomotive_potential(f)

Return the ponderomotive potential $U_p$, which is the cycle-average quiver energy of a free electron in an electromagnetic field f. It is given by

\[U_p = \frac{e^2E_0^2}{4m\omega^2}=\frac{2e^2}{c\varepsilon_0m}\times\frac{I}{4\omega^2},\]

or, in atomic units,

\[U_p = \frac{I}{4\omega^2}.\]

ElectricFields.keldysh — Function

keldysh(f, Iₚ)

The Keldysh parameter relates the strength of a dynamic electric field to that of the binding potential of an atom. It is given by

\[\gamma = \sqrt{\frac{I_p}{2U_p}},\]

where $I_p$ is the ionization potential of the atom and $U_p$ is the ponderomotive potential of the dynamic field.

ElectricFields.free_oscillation_amplitude — Function

free_oscillation_amplitude(F)

Compute the free oscillation amplitude of an electric field F, i.e. the mean excursion length during one cycle of the field, defined as

\[\alpha \defd \frac{F}{\omega^2}\]

where F is the peak amplitude, i.e. this is defined for one cycle of a monochrome field.