Reference

Internals

ElectricFields.calc_params! — Function

calc_params!(field_params)

This function performs the calculation of different quantities from the information provided.

ElectricFields.@namespace! — Macro

@namespace!(exprs, params)

This macro uses the dictionary params as a "namespace", i.e. all symbols are assumed to be keys in this dictionary. We need to escape the generated expression tree to modify in the scope of the caller, not the global scope.

ElectricFields.gaussian_common! — Function

gaussian_common!(field_params, carrier[; Tmax_rounder, verbosity])

Compute parameters common to Gaussian envelopes, i.e. GaussianEnvelope and TruncatedGaussianEnvelope; most importantly, given an intensity FWHM or σ, we need to figure out the coefficient $\alpha$ for the envelope of the vector potential $\exp(-\alpha t^2)$, such that the electric field amplitude and intensity have the desired durations. The optional function Tmax_rounder determines if the envelop should be extended to encompass an integer amount of cycles of the carrier (default).

ElectricFields.test_field_parameters — Function

test_field_parameters(field_params, set)

This function ensures that one and only one of "competing" quantities is specified.

ElectricFields.convolution — Function

convolution(f̂::Function, dc::DiracComb, ω)

Evaluate the convolution between the function f̂ (assumed to be the Fourier transform of a function f) with the DiracComb dc. This is used to implement the Fourier transform of a function product $f(t)g(t)$, where $g(t)$ is a sum of monochromatic waves:

\[f(t)g(t) \rightsquigarrow \frac{1}{\sqrt{2\pi}} (\hat{f}\star\hat{g})(\omega)\]

ElectricFields.DiracComb — Type

DiracComb(frequencies)

Represents a Dirac frequency comb, where frequencies is a vector of Tuples: (ωᵢ,cᵢ), representing a frequency and an amplitude:

\[\mathrm{DC}(\omega) = \sum_i c_i \delta(\omega-\omega_i).\]

Index

ElectricFields.BK7

ElectricFields.Calcite

ElectricFields.KTP

ElectricFields.Quartz

ElectricFields.SiO₂

ElectricFields.AbstractCarrier

ElectricFields.AbstractWindow

ElectricFields.ApodizedField

ElectricFields.BSpline

ElectricFields.BSplineField

ElectricFields.Blackman

ElectricFields.BlackmanExact

ElectricFields.BlackmanHarris

ElectricFields.BlackmanNuttall

ElectricFields.CascadedDispersiveElement

ElectricFields.Chirp

ElectricFields.ConstantField

ElectricFields.Cos²Envelope

ElectricFields.Crystal

ElectricFields.Crystal

ElectricFields.DelayedField

ElectricFields.DiracComb

ElectricFields.DispersedField

ElectricFields.DispersiveElement

ElectricFields.EllipticalCarrier

ElectricFields.FixedCarrier

ElectricFields.GaussianEnvelope

ElectricFields.Hamming

ElectricFields.Hann

ElectricFields.IsotropicMedium

ElectricFields.IsotropicMedium

ElectricFields.Kaiser

ElectricFields.LinearCarrier

ElectricFields.LinearField

ElectricFields.LinearTransverseCarrier

ElectricFields.LinearTransverseField

ElectricFields.Medium

ElectricFields.NegatedField

ElectricFields.Nuttall

ElectricFields.PaddedField

ElectricFields.PhaseShift

ElectricFields.Ramp

ElectricFields.Rect

ElectricFields.Sellmeier

ElectricFields.SumField

ElectricFields.TransverseCarrier

ElectricFields.TransverseField

ElectricFields.TrapezoidalEnvelope

ElectricFields.TruncatedGaussianEnvelope

ElectricFields.WindowedField

ElectricFields.WrappedField

AbstractFFTs.fft

AbstractFFTs.irfft

AbstractFFTs.rfft

ElectricFields.amplitude

ElectricFields.calc_params!

ElectricFields.carrier

ElectricFields.chirp

ElectricFields.continuity

ElectricFields.convolution

ElectricFields.delay

ElectricFields.dimensions

ElectricFields.duration

ElectricFields.envelope

ElectricFields.fft_vector_potential

ElectricFields.fftω

ElectricFields.field_amplitude

ElectricFields.field_amplitude_spectrum

ElectricFields.field_envelope

ElectricFields.find_time_span

ElectricFields.fluence

ElectricFields.free_oscillation_amplitude

ElectricFields.frequency

ElectricFields.gaussian_common!

ElectricFields.instantaneous_intensity

ElectricFields.intensity

ElectricFields.intensity

ElectricFields.keldysh

ElectricFields.max_frequency

ElectricFields.nfft

ElectricFields.nfft_vector_potential

ElectricFields.period

ElectricFields.phase

ElectricFields.phase_shift

ElectricFields.phase_shift

ElectricFields.photon_energy

ElectricFields.polarization

ElectricFields.ponderomotive_potential

ElectricFields.rfft_vector_potential

ElectricFields.rfftω

ElectricFields.span

ElectricFields.span

ElectricFields.spectrum

ElectricFields.steps

ElectricFields.test_field_parameters

ElectricFields.time_bandwidth_product

ElectricFields.timeaxis

ElectricFields.vector_potential

ElectricFields.vector_potential_spectrum

ElectricFields.wavelength

ElectricFields.wavenumber

ElectricFields.@field

ElectricFields.@namespace!