Carriers

ElectricFields.AbstractCarrier — Type

AbstractCarrier

Any carrier, of any dimensionality.

ElectricFields.LinearCarrier — Type

LinearCarrier <: AbstractCarrier

Any carrier which is one-dimensional, i.e. linearly polarized (by convention along $z$).

ElectricFields.FixedCarrier — Type

FixedCarrier <: LinearCarrier

The carrier is fixed in the sense that the instantaneous frequency is constant throughout the pulse, i.e. no chirp/dispersion:

\[\Im\{\exp[\im(\omega t + \phi)]\} = \sin(\omega t + \phi).\]

ElectricFields.TransverseCarrier — Type

TransverseCarrier <: AbstractCarrier

Any carrier which is two-dimensional, i.e. polarized in the plane perpendicular to the direction of propagation.

ElectricFields.LinearTransverseCarrier — Type

LinearTransverseCarrier

The carrier is identical to FixedCarrier, but explicitly in 3d, with the polarization taken along the $z$ axis:

\[\vec{C}(t) = \bmat{0\\0\\\sin(\omega t + \phi)}.\]

ElectricFields.EllipticalCarrier — Type

EllipticalCarrier

An elliptically polarized field (of which circularly and linearly polarized are special cases) is given in the canonical coordinate system as

\[\vec{C}(t) = \frac{1}{\sqrt{1+\xi^2}} \bmat{\xi\cos\theta\\0\\\sin\theta}, \quad \theta \defd \omega t + \phi,\]

i.e. with the principal axis of the ellipse along $z$.