Reference

ElectricFields.AbstractCarrier

ElectricFields.ConstantField

ElectricFields.Cos²Envelope

ElectricFields.DelayedField

ElectricFields.EllipticalCarrier

ElectricFields.FixedCarrier

ElectricFields.GaussianEnvelope

ElectricFields.LinearCarrier

ElectricFields.LinearField

ElectricFields.LinearTransverseCarrier

ElectricFields.NegatedField

ElectricFields.PaddedField

ElectricFields.SumField

ElectricFields.TransverseCarrier

ElectricFields.TransverseField

ElectricFields.TrapezoidalEnvelope

ElectricFields.TruncatedGaussianEnvelope

ElectricFields.WindowedField

ElectricFields.WrappedField

ElectricFields.amplitude

ElectricFields.calc_params!

ElectricFields.carrier

ElectricFields.continuity

ElectricFields.continuity

ElectricFields.dimensions

ElectricFields.duration

ElectricFields.duration

ElectricFields.envelope

ElectricFields.field_amplitude

ElectricFields.field_amplitude

ElectricFields.field_envelope

ElectricFields.fluence

ElectricFields.fluence

ElectricFields.free_oscillation_amplitude

ElectricFields.free_oscillation_amplitude

ElectricFields.frequency

ElectricFields.frequency

ElectricFields.gaussian_common!

ElectricFields.instantaneous_intensity

ElectricFields.intensity

ElectricFields.intensity

ElectricFields.keldysh

ElectricFields.keldysh

ElectricFields.max_frequency

ElectricFields.max_frequency

ElectricFields.period

ElectricFields.period

ElectricFields.phase

ElectricFields.phase_shift

ElectricFields.photon_energy

ElectricFields.photon_energy

ElectricFields.polarization

ElectricFields.ponderomotive_potential

ElectricFields.ponderomotive_potential

ElectricFields.span

ElectricFields.spectrum

ElectricFields.steps

ElectricFields.test_field_parameters

ElectricFields.timeaxis

ElectricFields.timeaxis

ElectricFields.vector_potential

ElectricFields.wavelength

ElectricFields.wavelength

ElectricFields.wavenumber

ElectricFields.wavenumber

ElectricFields.@field

ElectricFields.@namespace!

ElectricFields.calc_params! — Method

calc_params!(field_params)

This function performs the calculation of different quantities from the information provided.

ElectricFields.continuity — Method

continuity(f)

Return the pulse continuity, i.e. differentiability, of the field f.

ElectricFields.duration — Method

duration(f)

Return the pulse duration $\tau$ of the field f.

ElectricFields.field_amplitude — Method

field_amplitude(f, a, b)

Compute the time integral of the field amplitude according to

\[\int_a^b\diff{t}F(t) = -[A(b) - A(a)],\]

where $A(t)$ is the vector_potential.

ElectricFields.field_amplitude — Method

field_amplitude(f, t)

Compute the field amplitude from the vector_potential $A(t)$ using automatic differentiation according to

\[F(t) = -\partial_t A(t).\]

ElectricFields.field_envelope — Method

field_envelope(f, t)

Compute the field amplitude envelope as the square root of field_envelope.

ElectricFields.fluence — Method

fluence(f)

Compute the fluence of the field f, i.e.

\[\frac{1}{\hbar\omega} \int\diff{t} I(t),\]

where $I(t)$ is the intensity envelope of the pulse.

ElectricFields.free_oscillation_amplitude — Method

free_oscillation_amplitude(F)

Compute the free oscillation amplitude of an electric field F, i.e. the mean excursion length during one cycle of the field, defined as

\[\alpha \defd \frac{F}{\omega^2}\]

where F is the peak amplitude, i.e. this is defined for one cycle of a monochrome field.

ElectricFields.frequency — Method

frequency(f)

Return the carrier frequency $f$ of the field f.

ElectricFields.gaussian_common! — Method

gaussian_common!(field_params, carrier[; Tmax_rounder, verbosity])

Compute parameters common to Gaussian envelopes, i.e. GaussianEnvelope and TruncatedGaussianEnvelope; most importantly, given an intensity FWHM or σ, we need to figure out the coefficient $\alpha$ for the envelope of the vector potential $\exp(-\alpha t^2)$, such that the electric field amplitude and intensity have the desired durations. The optional function Tmax_rounder determines if the envelop should be extended to encompass an integer amount of cycles of the carrier (default).

ElectricFields.instantaneous_intensity — Method

instantaneous_intensity(f, t)

Compute the instantaneous intensity of the field f as

\[I_i(t) = \abs{F(t)}^2 = \abs{-\partial_t A(t)}^2.\]

ElectricFields.intensity — Method

intensity(f, t)

Compute the intensity envelope of the field f at time t by applying phase_shift to the carrier and looking for the maximal instantaneous_intensity.

ElectricFields.keldysh — Method

keldysh(f, Iₚ)

The Keldysh parameter relates the strength of a dynamic electric field to that of the binding potential of an atom. It is given by

\[\gamma = \sqrt{\frac{I_p}{2U_p}},\]

where $I_p$ is the ionization potential of the atom and $U_p$ is the ponderomotive potential of the dynamic field.

ElectricFields.max_frequency — Method

max_frequency(f)

Return the maximum carrier frequency $f$ in case of a composite field f.

ElectricFields.period — Method

period(f)

Return the carrier period time $T$ of f.

ElectricFields.phase_shift — Function

phase_shift(f, δϕ)

Return a new field whose carrier has been phase-shifted by δϕ.

ElectricFields.photon_energy — Method

photon_energy(f)

Return the carrier photon energy $\hbar\omega$ of the field f.

ElectricFields.ponderomotive_potential — Method

ponderomotive_potential(f)

Return the ponderomotive potential $U_p$, which is the cycle-average quiver energy of a free electron in an electromagnetic field f. It is given by

\[U_p = \frac{e^2E_0^2}{4m\omega^2}=\frac{2e^2}{c\varepsilon_0m}\times\frac{I}{4\omega^2},\]

or, in atomic units,

\[U_p = \frac{I}{4\omega^2}.\]

ElectricFields.span — Method

span(f)

Return the end-points for the envelope of the field f.

ElectricFields.spectrum — Method

spectrum(env::GaussianEnvelope)

Gaussians belong to the Schwartz class, i.e. functions who, under Fourier transform, are mapped back to the same space. That is to say, the Fourier transform of a Gaussian is a Gaussian:

\[\exp(-\alpha t^2) \leftrightarrow \frac{1}{\sqrt{2\alpha}} \exp\left(-\frac{\omega^2}{4\alpha}\right).\]

Comparing with the above, we find that the spectral standard deviation

\[\Omega = \sqrt{2\alpha} = \frac{2\sqrt{\ln 2}}{\tau},\]

and the Gaussian function in the spectral domain is thus

\[E(\omega) = \frac{E_0\tau}{2\sqrt{\ln 2}} \exp\left[-\frac{(\omega\tau)^2}{8\ln2}\right].\]

ElectricFields.steps — Function

steps(f[, fs])

Return number of time steps for the field f, optionally supplying a sampling frequency.

We construct the time axis such that the highest frequency component of the field is resolved. By the Nyquist sampling theorem, we need minimum $f_s>2f_{\textrm{max}}$, but to be on the safe side, we use, as default, $f_s=100f_{\textrm{max}}$. This also makes plots nicer.

If fs<:Integer, the sampling frequency is computed as fs/T, i.e. fs steps per cycle.

ElectricFields.test_field_parameters — Method

test_field_parameters(field_params, set)

This function ensures that one and only one of "competing" quantities is specified.

ElectricFields.timeaxis — Function

timeaxis(f[, fs])

Construct a time axis for the field f, covering the span of the envelope in the number of steps given by the sample frequency fs.

ElectricFields.timeaxis — Method

timeaxis(f, (N,dt))

Alternate interface that constructs the time axis with N steps starting from first(span(f)), with intervals of dt.

ElectricFields.vector_potential — Function

vector_potential(f, t)

Compute the vector potential $\vec{A}(t)$ of the field f.

ElectricFields.wavelength — Method

wavelength(f)

Return the carrier wavelength $\lambda$ of f.

ElectricFields.wavenumber — Method

wavenumber(f)

Return the carrier wavenumber $\nu$ of the field f.

ElectricFields.@field — Macro

@field(name) do ... end

Frontend macro for creating an electric field and storing it in the variable name.

Example

julia> @field(F) do
           I₀ = 2.0
           T = 2.0
           σ = 3.0
           Tmax = 3.0
       end
Linearly polarized field with
  - I₀ = 2.0000e+00 au = 7.0188904e16 W cm⁻² =>
    - E₀ = 1.4142e+00 au = 727.2178 GV m⁻¹
    - A₀ = 0.4502 au
  – a Fixed carrier @ λ = 14.5033 nm (T = 48.3777 as, ω = 3.1416 Ha = 85.4871 eV)
  – and a Gaussian envelope of duration 170.8811 as (intensity FWHM; ±2.00σ)

ElectricFields.@namespace! — Macro

@namespace!(exprs, params)

This macro uses the dictionary params as a "namespace", i.e. all symbols are assumed to be keys in this dictionary. We need to escape the generated expression tree to modify in the scope of the caller, not the global scope.